# Minor optimizations