--- a/src/main.cpp
+++ b/src/main.cpp
@@ -4164,10 +4164,6 @@ bool CBlock::AddToBlockIndex(unsigned int nFile, unsigned int nBlockPos, const u
 
 bool CBlock::CheckBlock(std::string sCaller, int height1, int64_t Mint, bool fCheckPOW, bool fCheckMerkleRoot, bool fCheckSig, bool fLoadingIndex) const
 {
-
-    if (GetHash() == (fTestNet ? hashGenesisBlockTestNet : hashGenesisBlock))
-        return true;
-
     // These are checks that are independent of context
     // that can be verified before saving an orphan block.