Small optimizations to slightly improve the sync speeds. This imports 18.5k of testnet blocks from file around 1.5% faster. However, the first ~2k of those are scrypt blocks where the gain is proportionally smaller. If run on a larger block set the gain increase should be higher.

- Reduce the number of times block hash calculations are done in CheckBlock
- Avoid converting the CPID to its textual representation in GetLifetimeCPID

Original:
> 07/09/18 06:09:57 Loaded 18510 blocks from external file in 91614ms
> 07/09/18 06:11:49 Loaded 18510 blocks from external file in 91227ms

Opt:
> 07/09/18 06:07:32 Loaded 18510 blocks from external file in 90246ms
> 07/09/18 06:14:42 Loaded 18510 blocks from external file in 90663ms