@denravonska 2018-07-09
This PR changes the behavior. Now `str("") is equal to uint128()` which was not the case before.