@denravonska 2018-07-09
Looks like Im incorrect, but the behavior is still changed. With my changes it's not possible to use the CPID "INVESTOR" which done in a few places. I will have another look but might have to close the PR.