@TheCharlatan 2018-07-09
What is the difference between having the field empty and having "Investor"?