@denravonska 2018-07-09
I am not entirely sure to be honest. Let me do a full sync to see what's being passed to the function.