@denravonska 2018-07-09
Oh, it can never not be an invalid (!= 32 characters) CPID due to [this check](https://github.com/gridcoin-community/Gridcoin-Research/pull/1206/files#diff-7ec3c68a81efff79b6ca22ac1f1eabbaR5709), but the CPID returned from `GetCPID()` could either be `INVESTOR` or empty before whereas now it will be `0x0` for those two instances.