@denravonska 2018-07-11
I did a second run of this with a larger data set:

```
Original
org Loaded 784965 blocks from external file in 8801140ms
org Loaded 784965 blocks from external file in 8779245ms

Optimized
opt Loaded 784965 blocks from external file in 8429411ms
```

In this case the sync was ~4.1% faster with the optimizations. It's just a reduction of 6 minutes from 2.4 hours though.