@denravonska 2018-08-10
Completely removed the genesis block check in `CheckBlock`. This works on import, will try sync.